Update OrderBook sync logic
Web socket synchronization was in the trade tab itself, so when you go to the trade tab, it first displayed the spinner and then the real data. This is not professional, and it is now in detail page, so when you go to the trade tab, it displays the order book data directly, not loading.